requirements.txt
streamlit
google-generativeai==0.16.0
python-dotenv
PyPDF2
Chromadb
faiss-cpu
langchain==1.1.0
langchain-core==1.1.0
langchain-google-genai==2.0.10
langchain-text-splitters==1.0.0
langchain_community==0.4.1
langchain-classic==1.0.0


